+++ fdbclient/include/fdbclient/SystemData.h
@@ -91,6 +91,7 @@ void decodeKeyServersValue(RangeResult result,
                            UID& srcID,
                            UID& destID,
                            bool missingIsError = true);
+bool isSystemKey(KeyRef key);
 
 extern const KeyRangeRef auditRange;
 extern const KeyRef auditPrefix;
